fs/p9p/new: fully working ls with ufs
The repl and client session work fully with rminnich/go9p/ufs. The key to this
was proper support for directory serialization, which is subtle. The size field
for Rread was the incorrect size.

Several robustness improvements are also included. Issues with the timeouts and
missing error returns have been addressed.

Signed-off-by: Stephen J Day <stephen.day@docker.com>