fs/p9p/new: cleanup encoding/decoding
We take some time to cleanup the codec implementation. The need for the message
to implement the Message interface has been removed to allow flexibility in
what can land in the message field. Error types can be returned directly now.
The messages have been moved into a separate file. We also ensure that the
number of pointer types is reduced in deserialized messages, hopefully avoiding
unnecessary pointer traversal.

Signed-off-by: Stephen J Day <stephen.day@docker.com>