client: log messages with unknown tags
Several clients have experienced panics after receiving unknown tags.
The exact cause of these unknown messages is unknown. It is also not
clear whether an unknown tag should be fatal to the session.

This change provides more information in the panic message which we can
then use to better understand this condition.

Signed-off-by: Stephen J Day <stephen.day@docker.com>